// src/app/interfaces/vehicle-details.ts
export class VehicleDetails{
    id:             string;
    providerId:     string;
    provider:       string;
    eventName:      string;
    eventDate:      string;
    latitude:       number;
    longitude:      number;
    context: string;
  /*   context: {    
        eventType:string,
        speed: number,
        course:string,
        odometer:string
    }; */
    driverId:       string;
    driver:         string;
    assetId:        string;
    asset:          string;
    createdAt:      string;
}